+++ api/models/JobProcessSequenceTransaction.js
@@ -16,7 +16,7 @@ module.exports = {
     },
     quantity:{
       type:'number',
-      columnType: 'bigint',
+      columnType: 'integer',
     }
   },
 };
