+++ api/models/MaterialType.js
@@ -16,7 +16,7 @@ module.exports = {
     },
     status:{
     	type:'number',
-      columnType: 'bigint',
+      columnType: 'integer',
     }
   },
 };
